% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildElems.R
\name{buildElems}
\alias{buildElems}
\title{build multiple network elements by dataframe}
\usage{
buildElems(elems, type)
}
\arguments{
\item{elems}{value of elements consisted in dataframe}

\item{type}{Either 'Node' or 'Edge' [string]}
}
\description{
call buildNode or buildEdge function, note that only one function can be called
}
\examples{
nodes = buildElems( # will generate 5 nodes
 elems = data.frame(
   id = paste0('node',1:5),
   bgColor = "#FFFFFF",
   borderColor = "#48DBFB",
   borderWidth = 2,
   fontSize = 10,
   width = 60, height = 20, opacity = 1, stringsAsFactors = FALSE
 ), type = "Node")


}
\seealso{
buildNode(), buildEdge()
}
