% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCyJS.R
\name{renderShinyCyJS}
\alias{renderShinyCyJS}
\title{ShinyCyJS output}
\usage{
renderShinyCyJS(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expression that returns a list}

\item{env}{the environment in which to evaluate expr}

\item{quoted}{is expr a quoted expression (with quote())}
}
\description{
renders a cytoscape image for output
}
\seealso{
ShinyCyJSOutput()
}
