% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCyJS.R
\name{shinyCyJS}
\alias{shinyCyJS}
\title{cytoscape.js in shiny application}
\usage{
shinyCyJS(
  elements = list(),
  options = list(),
  layout = list(name = "cose"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{elements}{node and edge objects, it should be list of element.}

\item{options}{rendering / interaction options, can be created with buildIoption(), buildRoption()}

\item{layout}{list type layout, it must be contain name and other optional values}

\item{width}{canvas width.}

\item{height}{canvas height.}

\item{elementId}{id used to identify in application.}
}
\description{
generate canvas with given network element and options
}
