% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCyJS.R
\name{ShinyCyJSOutput}
\alias{ShinyCyJSOutput}
\title{create an cytoscape canvas element}
\usage{
ShinyCyJSOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read the canvas from}

\item{width}{canvas width}

\item{height}{canvas height}
}
\description{
render a renderShinyCyJS() within an application page.
}
\seealso{
renderShinyCyJS()
}
